pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "GradingWebUI"
version = "0.6.2"
description = "A Website for grading paper-based exams"
# readme = "README.md"
requires-python = ">=3.12"
license = {file = "LICENSE"}
authors = [
    {name = "Sam Ogden", email = "samuel.s.ogden@gmail.com"},
]
keywords = ["education", "quiz", "canvas", "lms", "assessment", "teaching", "exam", "testing"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "Topic :: Education :: Testing",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "python-dotenv==1.0.1",
    "PyYAML==6.0.1",
    "requests==2.32.4",
    "urllib3==2.6.3",
    "pandas==2.3.3",
    "fuzzywuzzy==0.18.0",
    "python-Levenshtein==0.27.3",
    "pymupdf==1.26.7",
    "scikit-learn==1.8.0",
    "fastapi==0.127.1",
    "uvicorn[standard]==0.40.0",
    "pydantic==2.12.5",
    "python-multipart==0.0.22",
    "pillow==12.1.1",
    "pyzbar==0.1.9",
    "opencv-python==4.12.0.88",
    "openai==2.14.0",
    "anthropic==0.75.0",
    "ollama==0.6.1",
    "QuizGenerator==0.22.2",
    "bcrypt==5.0.0",
    "canvasapi==3.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.2",
    "httpx==0.28.1",
    "pip-audit==2.9.0",
]

[project.urls]
Homepage = "https://github.com/OtterDen-Lab/GradingWebUI"
Documentation = "https://github.com/OtterDen-Lab/GradingWebUI/tree/main/documentation"
Repository = "https://github.com/OtterDen-Lab/GradingWebUI"
"Bug Tracker" = "https://github.com/OtterDen-Lab/GradingWebUI/issues"

[project.scripts]
# grade-assignments = "Autograder.grade_assignments:main"


[tool.hatch.build.targets.wheel]
packages = ["src/grading_web_ui", "lms_interface"]

[tool.hatch.build.targets.sdist]
include = ["src/grading_web_ui/**", "lms_interface/**", "pyproject.toml", "README.md", "LICENSE"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
filterwarnings = [
    "ignore:builtin type SwigPyPacked has no __module__ attribute:DeprecationWarning",
    "ignore:builtin type SwigPyObject has no __module__ attribute:DeprecationWarning",
    "ignore:builtin type swigvarlink has no __module__ attribute:DeprecationWarning",
]
